Extract ShapePainter to a separate class
