+++ lib/sleek_circular_slider.dart
@@ -1,15 +1,18 @@
 library sleek_circular_slider;
 
 export 'src/circular_slider.dart';
+export 'src/painters/background_painter.dart';
+export 'src/painters/circular_arc_painter.dart';
+export 'src/painters/current_value_painter.dart';
+export 'src/painters/progress_bar_painter.dart';
+export 'src/painters/shadow_painter.dart';
+export 'src/painters/shape_painter.dart';
+export 'src/parameters/circular_slider_callbacks.dart';
+export 'src/parameters/circular_slider_painters.dart';
 export 'src/parameters/circular_slider_settings.dart';
-export 'src/settings/circular_slider_features.dart';
+export 'src/parameters/circular_slider_values.dart';
 export 'src/settings/circular_slider_colors.dart';
+export 'src/settings/circular_slider_features.dart';
 export 'src/settings/circular_slider_shadow.dart';
 export 'src/settings/circular_slider_text.dart';
-export 'src/painters/background_painter.dart';
-export 'src/painters/shadow_painter.dart';
-export 'src/painters/progress_bar_painter.dart';
-export 'src/painters/current_value_painter.dart';
-export 'src/parameters/circular_slider_values.dart';
-export 'src/parameters/circular_slider_painters.dart';
-export 'src/parameters/circular_slider_callbacks.dart';
+export 'src/utilities/unit_conversions.dart';
