+++ lib/src/painters/circular_arc_painter.dart
@@ -1,17 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:sleek_circular_slider/sleek_circular_slider.dart';
+import 'package:sleek_circular_slider/src/painters/shape_painter.dart';
 import 'package:sleek_circular_slider/src/parameters/circular_slider_values.dart';
 import 'package:sleek_circular_slider/src/utilities/unit_conversions.dart';
-
-abstract class ShapePainter {
-  void drawShape(
-    Canvas canvas,
-    Size size,
-    Paint paint, {
-    bool ignoreAngle = false,
-    bool spinnerMode = false,
-  });
-}
 
 class CircularArcPainter extends ShapePainter {
   final CircularSliderSettings settings;
