+++ lib/src/painters/shape_painter.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+abstract class ShapePainter {
+  void drawShape(
+    Canvas canvas,
+    Size size,
+    Paint paint, {
+    bool ignoreAngle = false,
+    bool spinnerMode = false,
+  });
+}
